+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# Version 0.3.3 - My Side Chick is Kitted Out
+- Added Theme Support for `select` elements
+- Added Theme Support for `range` elements
+- Added Theme Support for Prose Mirror Menu Bar
+- Added Theme Support for Sidebar Tab - Combat
+- Added Theme Support for Sidebar Tab - Actors
+- Added Theme Support for Sidebar Tab - Items
+- Added Theme Support for Sidebar Tab - Journal
+- Added Theme Support for Sidebar Tab - Tables
+- Added Theme Support for Sidebar Tab - Cards
+- Added Theme Support for Sidebar Tab - Playlists
+- Added Theme Support for Sidebar Tab - Compendium
+- Added Support for Module Compatibility Checker
+
 # Version 0.3.2 - I'M ALIVE
 Overhauls Foundry Virtual Tabletops UI into a Classic Fantasy RPG Theme.
 - Completed Handcrafted Windows